move about page to app router, shared Footer in layout
