+++ apps/web/app/(base-org)/layout.tsx
@@ -1,4 +1,4 @@
-import { Nav } from 'apps/web/src/components/Layout/Nav/Nav';
+import Nav from 'apps/web/src/components/Layout/Nav/Nav';
 
 import type { Metadata } from 'next';
