+++ apps/web/src/components/CookieManagerButton/index.tsx
@@ -1,0 +1,21 @@
+'use client';
+
+import { useState, useCallback } from 'react';
+import { CookiePreferencesModal } from '@coinbase/cookie-banner';
+
+export function CookieManagerButton() {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleOpenModal = useCallback(() => setIsOpen(true), []);
+
+  const handleCloseModal = useCallback(() => setIsOpen(false), []);
+
+  return (
+    <>
+      <button type="button" className="appearance-none underline" onClick={handleOpenModal}>
+        Cookie Manager
+      </button>
+      {isOpen && <CookiePreferencesModal isOpen={isOpen} onClose={handleCloseModal} />}
+    </>
+  );
+}
