+++ apps/web/src/components/Layout/Footer/Footer.tsx
@@ -1,19 +1,11 @@
 'use client';
-import { useState, useCallback } from 'react';
 import Link from 'next/link';
-import { CookiePreferencesModal } from '@coinbase/cookie-banner';
-
 import { Icon } from '../../Icon/Icon';
 import { Logo } from '../../Logo/Logo';
 import { bridgeUrl, docsUrl } from 'apps/web/src/constants';
+import { CookieManagerButton } from 'apps/web/src/components/CookieManagerButton';
 
 export function Footer() {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const handleOpenModal = useCallback(() => setIsOpen(true), []);
-
-  const handleCloseModal = useCallback(() => setIsOpen(false), []);
-
   return (
     <footer className="z-10 mt-auto flex w-full justify-center bg-gray-90 lg:pb-64">
       <div className="flex w-full max-w-[1440px] flex-col justify-between p-8 lg:flex-row">
@@ -53,9 +45,7 @@ export function Footer() {
             <br />
             <Link href="/cookie-policy">Cookie Policy</Link>
             <br />
-            <button type="button" className="appearance-none underline" onClick={handleOpenModal}>
-              Cookie Manager
-            </button>
+            <CookieManagerButton />
           </p>
         </div>
         <div className="flex h-full flex-col gap-5 pt-24 lg:flex-row lg:gap-10 lg:pt-0">
@@ -117,7 +107,6 @@ export function Footer() {
           </div>
         </div>
       </div>
-      {isOpen && <CookiePreferencesModal isOpen={isOpen} onClose={handleCloseModal} />}
     </footer>
   );
 }
