+++ apps/web/src/components/StartBuildingOnBase/StartBuildingOnBase.tsx
@@ -3,7 +3,7 @@ import { docsUrl } from 'apps/web/src/constants';
 import Image from 'next/image';
 import Link from 'next/link';
 
-function ReadTheDocsButton() {
+async function ReadTheDocsButton() {
   return (
     <div className="w-[200px]">
       <Link href={docsUrl} target="_blank" rel="noreferrer noopener">
@@ -13,7 +13,7 @@ function ReadTheDocsButton() {
   );
 }
 
-export function StartBuildingOnBase() {
+export async function StartBuildingOnBase() {
   return (
     <section className="flex w-full max-w-[1440px] flex-row flex-wrap items-start justify-between bg-black px-8 py-8">
       <div className="pb-16 font-display text-3xl text-white md:text-5xl lg:text-6xl">
